reworked the ue stuff
