+++ lmplay/exp/embeddings/ue1_0/unified_embeddings_v1_3/model.py
@@ -5,8 +5,8 @@ from typing import Optional, Any
 from lmplay.modules import Block
 import tiktoken
 from lmplay.base.base_model import LMBase, LMRunnerBase
-from lmplay.exp.embeddings.unified_embeddings_v1_0.modules import UnifiedEmbedding, ConvertableEmbedding
-
+from lmplay.modules import ULinear
+from lmplay.exp.embeddings.modules import UnifiedEmbedding, ConvertableEmbedding
 
 class GPT2(LMBase):
   def __init__(self,
@@ -20,7 +20,7 @@ class GPT2(LMBase):
                front_embed_mul=8.0,
                for_train=True,
                keep_embed_on_cpu=False,
-               version="1.4",
+               version="1.3",
                **ignore):
     super().__init__(f"ue_v{version}_{front_embed_mul}_{num_blocks}L_{max_len}",
                      max_len=max_len,
@@ -31,24 +31,17 @@ class GPT2(LMBase):
                      ff_dropout=ff_dropout,
                      embed_dropout=embed_dropout,
                      front_embed_mul=front_embed_mul)
-    # same as 1.0 but adding another integration layer to the sacrificial network
+    #same as 1.0 but the linear is now a ULinear!
     keep_embed_on_cpu = for_train and keep_embed_on_cpu
     self.tokenizer = tiktoken.get_encoding("gpt2")
     vocab_size = self.tokenizer.n_vocab
 
     self.max_len = max_len
     if not for_train:
-      # this will convert any UE into a normal embedding. After this, if the model is saved, it can be re-loaded by the baseline model.
+      #this will convert any UE into a normal embedding. After this, if the model is saved, it can be re-loaded by the baseline model.
       self.tok_embed = ConvertableEmbedding(vocab_size, embed_dim, front_embed_mul)
     else:
-      self.tok_embed = UnifiedEmbedding(vocab_size,
-                                        embed_dim,
-                                        front_embed_mul,
-                                        keep_embed_on_cpu=keep_embed_on_cpu,
-                                        linear=nn.Linear,
-                                        gelu=False,
-                                        integration1_5=False,
-                                        integration2=False)
+      self.tok_embed = UnifiedEmbedding(vocab_size, embed_dim, front_embed_mul, keep_embed_on_cpu=keep_embed_on_cpu, linear=ULinear)
     self.pos_embed = nn.Parameter(torch.zeros(1, max_len, embed_dim))
     self.dropout = nn.Dropout(embed_dropout)
     self.blocks = nn.Sequential(*[Block(max_len,
@@ -59,11 +52,11 @@ class GPT2(LMBase):
     self.ln = nn.LayerNorm(embed_dim)
     self.fc = nn.Linear(embed_dim, vocab_size)
 
-  def forward(self, x: torch.Tensor, cache: Optional[list] = None):
+  def forward(self, x:torch.Tensor, cache:Optional[list] = None):
     seq_len = x.size(1)
     x_start = 0
     if cache is not None and len(cache) > 0:
-      # the is part of generate
+      #the is part of generate
       x_start = cache[0][0].size(1)
       seq_len += x_start
     assert seq_len <= self.max_len, "sequence longer than model capacity"
@@ -76,15 +69,14 @@ class GPT2(LMBase):
       x = block(x, cache=self._kv_cache(cache, i))
     x = self.ln(x)
     if cache is None:
-      # No cache then this is training and we need to decode the whole thing
+      #No cache then this is training and we need to decode the whole thing
       x = self.fc(x)
     else:
-      # Not training. We only care about the last one
-      x = self.fc(x[:, -1:, :])
+      #Not training. We only care about the last one
+      x = self.fc(x[:,-1:,:])
     if not cache is None:
       return x, cache
     return x
-
 
 class ModelRunner(LMRunnerBase):
   def __init__(self, max_batch_size=25):
@@ -97,7 +89,7 @@ class ModelRunner(LMRunnerBase):
                        strict=False,
                        **parameters) -> (LMBase, Any):
     model_args = model_args if model_args else dict()
-    for k, v in parameters.items():
+    for k,v in parameters.items():
       if k not in model_args:
         model_args[k] = v
     model = GPT2(for_train=self.for_train, **model_args)
